requirements.txt
requests==2.31.0
python-telegram-bot==21.0
python-dotenv==1.0.1
schedule==1.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
